\name{ribiosPCA-package}
\alias{ribiosPCA-package}
\alias{ribiosPCA}
\docType{package}
\title{
  Robust Principal Component Analysis in ribios
}
\description{
  Robust PCA in ribios
}
\details{
\tabular{ll}{
Package: \tab ribiosPCA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-12-20\cr
License: \tab What license is it under?\cr
}
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>

Maintainer: Jitao David Zhang <jitao_david.zhang@roche.com>
}
\keyword{ package }
%\seealso{
%}
